[refactor] 추가 코드 리뷰 반영
