+++ src/utils/parseScenarioDetail.ts
@@ -1,7 +1,7 @@
-import { formatDate } from "@/lib/utils/formatDate";
+import { formatDate } from "@/utils/formatDate";
 import type { ReadonlyURLSearchParams } from "next/navigation";
-import type { Scenario18Detail } from "../components/Scenario18";
-import { formatDateRange } from "@/lib/utils/formatDateRange";
+import type { Scenario18Detail } from "@/app/automaticpayment-scenario/components/Scenario18";
+import { formatDateRange } from "@/utils/formatDateRange";
 
 export function parseScenarioDetail(searchParams: URLSearchParams | ReadonlyURLSearchParams): Scenario18Detail {
   const status = searchParams.get("status") ?? "정상";
